Revert 236278 "Android: moves chrome/ to use long for JNI"
> Android: moves chrome/ to use long for JNI
> 
> BUG=317523
> 
> Review URL: https://codereview.chromium.org/67573003

TBR=bulach@chromium.org

Review URL: https://codereview.chromium.org/79463002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@236286 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/android/java/src/org/chromium/chrome/browser/NavigationPopup.java
@@ -49,7 +49,7 @@ public class NavigationPopup extends ListPopupWindow implements AdapterView.OnIt
 
     private final int mFaviconSize;
 
-    private long mNativeNavigationPopup;
+    private int mNativeNavigationPopup;
 
     /**
      * Constructs a new popup with the given history information.
@@ -234,7 +234,7 @@ public class NavigationPopup extends ListPopupWindow implements AdapterView.OnIt
 
     private static native String nativeGetHistoryUrl();
 
-    private native long nativeInit();
-    private native void nativeDestroy(long nativeNavigationPopup);
-    private native void nativeFetchFaviconForUrl(long nativeNavigationPopup, String url);
+    private native int nativeInit();
+    private native void nativeDestroy(int nativeNavigationPopup);
+    private native void nativeFetchFaviconForUrl(int nativeNavigationPopup, String url);
 }
