+++ chrome/android/java/src/org/chromium/chrome/browser/infobar/TranslateInfoBar.java
@@ -42,7 +42,7 @@ public class TranslateInfoBar extends TwoButtonInfoBar implements SubPanelListen
     private final boolean mShouldShowNeverBar;
     private final TranslateInfoBarDelegate mTranslateDelegate;
 
-    public TranslateInfoBar(long nativeInfoBarPtr, TranslateInfoBarDelegate delegate,
+    public TranslateInfoBar(int nativeInfoBarPtr, TranslateInfoBarDelegate delegate,
             int infoBarType, int sourceLanguageIndex, int targetLanguageIndex,
             boolean autoTranslatePair, boolean shouldShowNeverBar, String[] languages) {
         super(null, BACKGROUND_TYPE_INFO,
@@ -331,7 +331,7 @@ public class TranslateInfoBar extends TwoButtonInfoBar implements SubPanelListen
         return mInfoBarType;
     }
 
-    void changeInfoBarTypeAndNativePointer(int infoBarType, long newNativePointer) {
+    void changeInfoBarTypeAndNativePointer(int infoBarType, int newNativePointer) {
         if (infoBarType >= 0 && infoBarType < MAX_INFOBAR_INDEX) {
             mInfoBarType = infoBarType;
             replaceNativePointer(newNativePointer);
