+++ chrome/android/java/src/org/chromium/chrome/browser/profiles/Profile.java
@@ -12,9 +12,9 @@ import org.chromium.base.CalledByNative;
  */
 public class Profile {
 
-    private long mNativeProfileAndroid;
+    private int mNativeProfileAndroid;
 
-    private Profile(long nativeProfileAndroid) {
+    private Profile(int nativeProfileAndroid) {
         mNativeProfileAndroid = nativeProfileAndroid;
     }
 
@@ -23,7 +23,7 @@ public class Profile {
     }
 
     @CalledByNative
-    private static Profile create(long nativeProfileAndroid) {
+    private static Profile create(int nativeProfileAndroid) {
         return new Profile(nativeProfileAndroid);
     }
 
@@ -33,7 +33,7 @@ public class Profile {
     }
 
     @CalledByNative
-    private long getNativePointer() {
+    private int getNativePointer() {
         return mNativeProfileAndroid;
     }
