+++ chrome/browser/android/new_tab_page_prefs.cc
@@ -17,10 +17,10 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 using base::android::ConvertJavaStringToUTF8;
 
-static jlong Init(JNIEnv* env, jclass clazz, jobject profile) {
+static jint Init(JNIEnv* env, jclass clazz, jobject profile) {
   NewTabPagePrefs* new_tab_page_prefs =
       new NewTabPagePrefs(ProfileAndroid::FromProfileAndroid(profile));
-  return reinterpret_cast<intptr_t>(new_tab_page_prefs);
+  return reinterpret_cast<jint>(new_tab_page_prefs);
 }
 
 NewTabPagePrefs::NewTabPagePrefs(Profile* profile)
