+++ chrome/browser/ui/android/infobars/auto_login_infobar_delegate_android.cc
@@ -47,7 +47,7 @@ bool AutoLoginInfoBarDelegateAndroid::AttachAccount(
   DCHECK(delegate.obj());
   user_ = base::android::ConvertJavaStringToUTF8(
       Java_AutoLoginDelegate_initializeAccount(
-          env, delegate.obj(), reinterpret_cast<intptr_t>(this), jrealm.obj(),
+          env, delegate.obj(), reinterpret_cast<jint>(this), jrealm.obj(),
           jaccount.obj(), jargs.obj()));
   return !user_.empty();
 }
@@ -64,7 +64,7 @@ bool AutoLoginInfoBarDelegateAndroid::Accept() {
   DCHECK(delegate.obj());
 
   Java_AutoLoginDelegate_logIn(env, delegate.obj(),
-                               reinterpret_cast<intptr_t>(this));
+                               reinterpret_cast<jint>(this));
 
   // Do not close the infobar on accept, it will be closed as part
   // of the log in callback.
@@ -77,7 +77,7 @@ bool AutoLoginInfoBarDelegateAndroid::Cancel() {
       weak_java_auto_login_delegate_.get(env);
   DCHECK(delegate.obj());
   Java_AutoLoginDelegate_cancelLogIn(env, delegate.obj(),
-                                     reinterpret_cast<intptr_t>(this));
+                                     reinterpret_cast<jint>(this));
   return true;
 }
